+++ packages/serinus/bin/serinus.dart
@@ -94,25 +94,27 @@ class AnotherController extends Controller {
     //   )
     // );
     on(
-      Route.all('/'),
+      Route.all(
+        '/',
+        pipes: {
+          BodySchemaValidationPipe(object({}).passthrough().list()),
+          TransformPipe((context) async {
+            final requestContext = context.switchToHttp();
+            requestContext.body = JsonBody.fromJson([
+              for (var item in requestContext.bodyAs<JsonList>().value)
+                item..['data'] = 'hello!',
+            ]);
+          }),
+          TransformPipe((context) async {
+            final argsHost = context.argumentsHost;
+            if (argsHost is! HttpArgumentsHost) {
+              return;
+            }
+            argsHost.request.query['transform'] = 'true';
+          }),
+        },
+      ),
       _fallback,
-      pipes: [
-        BodySchemaValidationPipe(object({}).passthrough().list()),
-        TransformPipe((context) async {
-          final requestContext = context.switchToHttp();
-          requestContext.body = JsonBody.fromJson([
-            for (var item in requestContext.bodyAs<JsonList>().value)
-              item..['data'] = 'hello!',
-          ]);
-        }),
-        TransformPipe((context) async {
-          final argsHost = context.argumentsHost;
-          if (argsHost is! HttpArgumentsHost) {
-            return;
-          }
-          argsHost.request.query['transform'] = 'true';
-        }),
-      ],
     );
   }
 
@@ -264,6 +266,24 @@ class AppController extends Controller with SseController {
       logger.info('Emitted event to TCP transport for pattern "*"');
       return 'Hello world!';
     });
+    on(
+      Route.post(
+        '/',
+        pipes: {
+          BodySchemaValidationPipe(
+            object({
+              'name': string().min(3).max(50),
+              'birthdate': string().dateTime(),
+            }),
+          ),
+        },
+      ),
+      (RequestContext context) {
+        final body = context.body;
+        logger.info('Body: $body');
+        return body;
+      },
+    );
     onSse(Route.get('/sse'), (SseContext context) async* {
       yield 'Hello';
       await Future.delayed(Duration(seconds: 3));
@@ -279,6 +299,8 @@ void main(List<String> arguments) async {
     port: 3002,
     logger: ConsoleLogger(prefix: 'Serinus New Logger'),
   );
+  application.globalPrefix = 'api';
+  application.versioning = VersioningOptions(type: VersioningType.uri);
   application.enableShutdownHooks();
   // application.trace(ServerTimingTracer());
   await application.serve();