+++ packages/serinus/lib/src/handlers/handlers.dart
@@ -1,3 +1,0 @@
-// export 'handler.dart';
-// export 'request_handler.dart';
-// export 'websocket_handler.dart';