+++ packages/serinus_swagger/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0-rc.1
+
+- updated serinus to 2.0.0-rc.1
+
 ## 1.0.3
 
 - Updated `serinus_swagger` to be compatible with `serinus` version `1.x.x`.